requirements.txt
fastapi
uvicorn
numpy
opencv-python
requests
python-multipart
scipy
Pillow
rknn-toolkit2==2.3.2